+++ app.py
@@ -6,7 +6,10 @@ from pprint import pprint
 
 from framework.framework import Framework
 from framework.views import BaseViewLayout, BaseViewSection
+from framework.logger import LoggerFabric, Log, LoggerLevel
+
 from urls import app_urls
+from logs import loggers, handlers
 from category import Category
 from course import Course
 
@@ -15,6 +18,11 @@ import settings
 # Прочие константы
 EMAIL_ADMIN = "nekidoz@yandex.ru"
 
+logger = LoggerFabric(handlers=handlers, loggers=loggers)
+log_d = Log(settings.LOGGER_DEBUG, LoggerLevel.DEBUG, 'app')
+log_r = Log(settings.LOGGER_RUNTIME, LoggerLevel.INFO, 'app')
+log_d("Test debug message")
+log_r("Test runtime message")
 app = Framework(app_urls, use_admin=True)                   # Create application object
 if not (Category.register() and Course.register()):         # Initialize persistence data
     print("Failed to init data sources")